% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_mae.seg.R
\name{annotate.gr_seg}
\alias{annotate.gr_seg}
\title{annotate_cnv.seg}
\usage{
annotate.gr_seg(gr_seg, lst_gen)
}
\arguments{
\item{gencode_fl}{}
}
\description{
incase of multiple values per gene, get the more extreme value
}
