% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R, R/annotate2.R
\name{annotate}
\alias{annotate}
\alias{annotate}
\title{Generates a cmdline, when invoked annotates files using annovar}
\usage{
annotate(files, samplename = opts_flow$get("samplename"), outfile,
  build = "hg19", annovar_dir = opts_flow$get("annovar_dir"),
  annotate_func = "ultraseq::annotate_mutect")

annotate(files, samplename = opts_flow$get("samplename"), outfile,
  build = "hg19", annovar_dir = opts_flow$get("annovar_dir"),
  annotate_func = "ultraseq::annotate_mutect")
}
\arguments{
\item{samplename}{name of the sample, to be added to the flowmat}

\item{outfile}{name of the final merged file.}

\item{x}{files to be annotates}

\item{x}{files to be annotates}

\item{samplename}{name of the sample, to be added to the flowmat}

\item{outfile}{name of the final merged file.}
}
\description{
Generates a cmdline, when invoked annotates files using annovar

Generates a cmdline, when invoked annotates files using annovar
}
