% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate2.R
\name{annotate_variants}
\alias{annotate_variants}
\title{annotate variants}
\usage{
annotate_variants(x, outfile, execute = TRUE,
  annovar_path = "/scratch/iacs/apps/annovar/latest",
  db_path = "/scratch/iacs/apps/annovar/humandb",
  annovar_opts = "-otherinfo -nastring . -remove -buildver",
  build = "hg19", protocols_g, protocols_f, chr = "contig",
  start = "position", end = "position", ref = "ref_allele",
  alt = "alt_allele", use_uuid = TRUE)
}
\arguments{
\item{x}{A data.frame or a file with an extension recognized by read_sheet}

\item{execute}{logical, run annovar dry-run}

\item{annovar_path}{Path to annovar executable}

\item{db_path}{path to annovar database}

\item{build}{build version of the input files}

\item{protocols_g}{region-type annotation databases, see Annovar's website for details}

\item{protocols_f}{filter-type annotation databases, see Annovar's website for details}

\item{chr}{column which represents the chromosome}

\item{start}{column representing start location}

\item{end}{column representing end location}

\item{ref}{column representing reference allele}

\item{alt}{column representing alternate allele}

\item{out_path}{output path}

\item{out_file}{output file}
}
\description{
annotate variants
}
\details{
Important default arguments:

\code{-otherinfo -nastring . -remove -buildver}
}
