% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate2.R
\name{annotate_vcf}
\alias{annotate_vcf}
\title{Not tested}
\usage{
annotate_vcf(
  x,
  oprefix,
  odir,
  build,
  use_uuid,
  annovar_path,
  protocols_f,
  protocols_g,
  ...
)
}
\description{
Not tested
}
