% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_genomes.R
\name{avail_genomes}
\alias{avail_genomes}
\title{Use Illumina's iGenomes and get a list of available genomes}
\usage{
avail_genomes(
  species,
  src,
  build,
  from = "igenomes",
  verbose = FALSE,
  base_url = "ftp://igenome:G3nom3s4u@ussd-ftp.illumina.com",
  ...
)
}
\arguments{
\item{species}{species}

\item{src}{source}

\item{build}{build}

\item{from}{from}

\item{verbose}{logical, specifying verboseness}

\item{base_url}{base url to be used}

\item{...}{not used}
}
\description{
additional details
}
\examples{
gen = avail_genomes(species = 'Homo_Sapiens', from = 'igenomes')
}
