% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_readcount.R
\name{bam_readcount.parse}
\alias{bam_readcount.parse}
\title{parse read count}
\usage{
bam_readcount.parse(x, samplename, bed, type = c("snv", "indel"))
}
\arguments{
\item{x}{something}

\item{samplename}{something}

\item{bed}{something}
}
\description{
parse read count
}
