% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_readcount.R
\name{bam_readcount}
\alias{bam_readcount}
\title{Run BAM readcount}
\usage{
bam_readcount(
  bam,
  samplename,
  bed,
  outfile,
  fa_fl = "~/ref/human/b37/fastas/Homo_sapiens_assembly19.fasta",
  bamreadcount_exe = "~/apps/conda/3/envs/bamreadcount/bin/bam-readcount"
)
}
\arguments{
\item{bam}{something}

\item{samplename}{something}

\item{bed}{something}

\item{outfile}{something}

\item{fa_fl}{something}

\item{bamreadcount_exe}{something}
}
\description{
Run BAM readcount
}
\details{
should be 1-based:

The list of regions should be formatted as chromosome start and end. Each field should be tab separated and coordinates should be 1-based.
}
