% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_readcount_py.R
\name{bam_readcount_helper}
\alias{bam_readcount_helper}
\title{bam_readcount_helper}
\usage{
bam_readcount_helper(
  samplename,
  vcf,
  bam,
  tumor_seq_sample_id,
  ref_fasta = opts_flow$get("ref_fasta"),
  ensembl_vep_setup = opts_flow$get("ensembl_vep_setup"),
  vt_exe = opts_flow$get("vt_exe"),
  bam_readcount_helper_py = opts_flow$get("bam_readcount_helper_py")
)
}
\arguments{
\item{samplename}{}

\item{vcf}{}

\item{bam}{}

\item{tumor_seq_sample_id}{}

\item{ref_fasta}{}

\item{ensembl_vep_setup}{}

\item{vt_exe}{}

\item{bam_readcount_helper_py}{}
}
\value{

}
\description{
bam_readcount_helper
}
