% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_readcount.R
\name{bam_readcount_r}
\alias{bam_readcount_r}
\title{bam_readcount_r}
\usage{
bam_readcount_r(trk, execute = F)
}
\description{
# read add mutect files, filter and create bed
}
\details{
quick R func, for somatic variants
}
