% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_readcount.R
\name{bam_readcount_r.ssm}
\alias{bam_readcount_r.ssm}
\title{bam_readcount_r.ssm}
\usage{
bam_readcount_r.ssm(
  trk,
  conf,
  col_fl = "pcgr_combvcf_rds",
  col_bam = "bam",
  col_samp = "name",
  execute = F,
  bamreadcount_exe = "~/apps/conda/3/envs/bamreadcount/bin/bam-readcount",
  force_redo = F,
  fa_fl = "~/ref/human/b37/fastas/Homo_sapiens_assembly19.fasta"
)
}
\arguments{
\item{trk}{}

\item{conf}{}

\item{col_fl}{}

\item{col_bam}{}

\item{col_samp}{}

\item{execute}{}

\item{bamreadcount_exe}{}

\item{force_redo}{}

\item{fa_fl}{}
}
\description{
bam_readcount_r.ssm
}
