% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_set.R
\name{bam_set}
\alias{bam_set}
\title{.depreciated}
\usage{
bam_set(bam, outprefix, ref.fasta, interval_split, split_by = c("none",
  "chr_len", "interval_split"))
}
\arguments{
\item{bam}{}

\item{outfile}{}

\item{ref_fasta_path}{}

\item{split_by_chr}{}
}
\description{
.depreciated
}
