% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipseq.R
\name{chipseq}
\alias{chipseq}
\title{A complete chipseq pipeline from fastq to MACS and Scripture}
\usage{
chipseq(
  fqs,
  samplename = "sample_name",
  out_path,
  deffile = system.file("data/chipseq.def", package = "SaturnV"),
  bowtie_exe = "/risapps/rhel6/bowtie/1.1.2/bowtie",
  java_exe = "/risapps/noarch/jdk/jdk1.8.0_45/bin/java",
  igvtools_exe = "/rsrch2/genomic_med/krai/apps/igvtools/default/igvtools.jar",
  scripture_exe = "/rsrch2/genomic_med/krai/apps/scripture/scripture-beta2.jar",
  macs14_exe = "macs14",
  bowtie_index = "/rsrch2/genomic_med/krai/annot/indexes/bowtie/hg19",
  hg19table = "/rsrch2/genomic_med/krai/annot/hg19.table",
  chr_prefix = "",
  peak_fdr = "0.05"
)
}
\arguments{
\item{fqs}{a character vector of paths to fastqs for a sample}

\item{samplename}{name of this sample (a character vector, length one.)}
}
\description{
A complete chipseq pipeline from fastq to MACS and Scripture
}
