% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwa.R
\name{chk_fq}
\alias{chk_fq}
\title{chk fq:}
\usage{
chk_fq(fqs1, fqs2)
}
\arguments{
\item{fqs1}{a vector of file paths}

\item{fqs2}{a vector of file paths}
}
\description{
Files should have same extension. If paired, both should be of the same length
}
