% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_results.R
\name{copy_data_base}
\alias{copy_data_base}
\title{copy_data_base}
\usage{
copy_data_base(fls, out_dir, transfer = FALSE)
}
\arguments{
\item{fls}{}

\item{out_dir}{}

\item{transfer}{}
}
\description{
copy_data_base
}
