% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_paired.R
\name{create_paired_samplesheet}
\alias{create_paired_samplesheet}
\title{create_paired_samplesheet}
\usage{
create_paired_samplesheet(x, fqmat, sampbam, refbam, outfile, db_sampleid = 0,
  db_refid = 0, out_prefix, bampath, project = "project",
  subproject = "subproject", normal_bam_01)
}
\arguments{
\item{normal_bam_01}{}
}
\description{
create_paired_samplesheet
}
