% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sheet.R
\name{create_sample_sheet}
\alias{create_sample_sheet}
\title{Creates a sample sheet file names in the provided folder}
\usage{
create_sample_sheet(
  path,
  project,
  subproject,
  runid,
  format,
  fix.names = FALSE,
  fix.names.char = "-",
  out_sep = c("\\t", ","),
  include.undetermined = FALSE,
  pattern = "fastq.gz|fq.gz|fastq|fq",
  outfile
)
}
\arguments{
\item{path}{path to the fastq files}

\item{project}{name of the project.  \emph{optional}}

\item{subproject}{name of the subproject \emph{optional}}

\item{runid}{name of the flowcell this data is from \emph{optional}}

\item{format}{the format for names of fastq files, we have defaults for CASAVA and miSeq}

\item{fix.names}{change the sample names such that they are acceptable as column names for R}

\item{pattern}{extensions this function will look for \emph{optional}}

\item{outfile}{name of the output csv files \emph{optional}}
}
\description{
This function would check for files ending in fastq.gz, fq.gz, fastq, fq.
}
\examples{
\dontrun{

create_sample_mat(levelipath)

}
}
\keyword{casava}
\keyword{fastq}
\keyword{samplesheet}
