% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto.R
\name{create_trk.kallisto}
\alias{create_trk.kallisto}
\title{create a tracking sheet for kallisto}
\usage{
create_trk.kallisto(
  trk,
  col_sequencing_id = "sample_id",
 
    kallisto_path = "~/.rsrch1/iacs/iacs_dep/sseth/data/runs/20180313_CT_RNA/kallisto/flowname-20180716-17-04-03-5WlTgRNo"
)
}
\arguments{
\item{trk}{something}

\item{base_out}{something}

\item{project_id}{something}

\item{sequencing_id}{something}
}
\description{
create a tracking sheet for kallisto
}
