% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_pindel.R
\name{.gm_pindel.read}
\alias{.gm_pindel.read}
\title{.gm_pindel.read}
\usage{
.gm_pindel.read(fl, fix_deletions = T)
}
\arguments{
\item{fl}{read a single pindel file}
}
\value{

}
\description{
.gm_pindel.read
}
