% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_vcf.R
\name{.mutect1_header}
\alias{.mutect1_header}
\title{##contig=<ID=20,length=62435964,assembly=B36,md5=f126cdf8a6e0c7f379d618ff66beb2da,species="Homo sapiens",taxonomy=x>',
##phasing=partial',
##INFO=<ID=DB,Number=0,Type=Flag,Description="dbSNP Membership">',
##INFO=<ID=TLOD,Number=1,Type=String,Description="CORE STATISTIC: Log of (likelihood tumor event is real / likelihood event is sequencing error )">',}
\usage{
.mutect1_header()
}
\description{
##contig=<ID=20,length=62435964,assembly=B36,md5=f126cdf8a6e0c7f379d618ff66beb2da,species="Homo sapiens",taxonomy=x>',
##phasing=partial',
##INFO=<ID=DB,Number=0,Type=Flag,Description="dbSNP Membership">',
##INFO=<ID=TLOD,Number=1,Type=String,Description="CORE STATISTIC: Log of (likelihood tumor event is real / likelihood event is sequencing error )">',
}
