% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets.R
\name{facets_r}
\alias{facets_r}
\title{facets_r}
\usage{
facets_r(rcfl, oprefix, pre_snp.nbhd = 250, pre_cval = 50,
  pre_hetscale = T, pre_gbuild = "hg19", pre_unmatched = F,
  pre_het.thresh = 0.25, proc_cval = 300, proc_min.nhet = 30)
}
\arguments{
\item{rcfl}{}

\item{oprefix}{}

\item{pre_snp.nbhd}{}

\item{pre_cval}{}

\item{pre_hetscale}{}

\item{pre_gbuild}{}

\item{pre_unmatched}{indicator of whether the normal sample is unmatched.
When this is TRUE hets are called using tumor reads only and logOR calculations are different. 
Use het.thresh = 0.1 or lower when TRUE.}

\item{proc_cval}{}

\item{proc_min.nhet}{}

\item{het.thresh}{vaf threshold to call a SNP heterozygous}
}
\description{
facets_r
}
