% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyclone.R
\name{facets_read}
\alias{facets_read}
\title{data frame of segment summaries pre and post clustering of segments. 
The columns are: 
‘chrom’ the chromosome to which the segment belongs; 
‘seg’ the segment number; 
‘num.mark’ the number of SNPs in the segment; 
‘nhet’ the number of SNPs that are deemed heterozygous; 
‘cnlr.median’ the median log-ratio of the segment; 
‘mafR’ the log-odds-ratio summary for the segment; 
‘segclust’ the segment cluster to which segment belongs; 
‘cnlr.median.clust’ the median log-ratio of the segment cluster; 
‘mafR.clust’ the log-odds-ratio summary for the segment cluster;
‘cf’ the cellular fraction of the segment; 
‘tcn’ the total copy number of the segment; 
‘lcn’ the minor copy number of the segment.}
\usage{
facets_read(segfl)
}
\description{
dataframe consisting of the columns of segmentation output as well as 
cellular fraction (cf), total (tcn) and lesser (lcn) copy number of each segment 
and their em counterpart (with .em suffix)
}
