% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc.R
\name{fastqc}
\alias{fastqc}
\title{Fastqc}
\usage{
fastqc(
  fqs,
  samplename = opts_flow$get("samplename"),
  fqpath,
  odir,
  fastqc_exe = opts_flow$get("fastqc_exe"),
  cpu_fastqc = opts_flow$get("cpu_fastqc"),
  casava = TRUE,
  fastqc_opts = opts_flow$get("fastqc_opts")
)
}
\arguments{
\item{fqs}{fqs}

\item{samplename}{}

\item{fqpath}{fqpath}

\item{odir}{odir}

\item{fastqc_exe}{fastqc_exe}

\item{cpu_fastqc}{something}

\item{casava}{something}

\item{fastqc_opts}{something}
}
\description{
Fastqc
}
