% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_genomes.R
\name{fetch_genomes}
\alias{fetch_genomes}
\title{fetch_genomes}
\usage{
fetch_genomes(genome_path = "~/flowr/genomes", verbose = FALSE, ...)
}
\arguments{
\item{genome_path}{output path to be used for downloading}

\item{verbose}{logical specifying verboseness}

\item{...}{passed onto \link{fetch_genomes}}
}
\description{
fetch_genomes
}
\examples{
\dontrun{
genomes_fetch(species = Homo_sapiens, src = 'NCBI', build = 'build37.2')
}
}
