% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_context.R
\name{get_context}
\alias{get_context}
\title{Title}
\usage{
get_context(x, bedtools_exe = "bedtools", tmpfile1 = tempfile(fileext =
  ".bed"), tmpfile2 = tempfile(fileext = ".bed"),
  reflib = "~/ref/human/fastas/Homo_sapiens_assembly19.fasta")
}
\arguments{
\item{x}{data.frame with chr start end}

\item{bedtools_exe}{}

\item{tmpfile1}{}

\item{tmpfile2}{}

\item{reflib}{}
}
\description{
Title
}
