% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_set.R
\name{get_fasta_chrs}
\alias{get_fasta_chrs}
\title{Read the associated dictionary file and return a list of chromosome names}
\usage{
get_fasta_chrs(fa = opts_flow$get("ref_fasta_path"), length = FALSE)
}
\arguments{
\item{fa}{a reference genome fasta file}

\item{length}{provide length of each chromosome as well [FALSE]}
}
\description{
Read the associated dictionary file and return a list of chromosome names
}
