% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_pindel.R
\name{gm_pindel.read}
\alias{gm_pindel.read}
\title{read pindel}
\usage{
gm_pindel.read(trk, col_fl = "pindel_fl", col_samp = "NAME", cores = 1)
}
\arguments{
\item{trk}{something}

\item{col_fl}{something}

\item{col_samp}{something}
}
\description{
read pindel
}
