% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_platypus.R
\name{gm_plat.read}
\alias{gm_plat.read}
\title{read mutect}
\usage{
gm_plat.read(trk, col_fl = "mutect", col_samp = "samplelbl1", cores = 1)
}
\arguments{
\item{trk}{something}

\item{col_fl}{something}

\item{col_samp}{something}
}
\description{
read mutect
}
