% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_caller.R
\name{haplotype_caller}
\alias{haplotype_caller}
\title{a wrapper around gatk haplotyper}
\usage{
haplotype_caller(
  x,
  samplename = opts_flow$get("samplename"),
  outfile,
  split_by_chr = TRUE,
  ref_fasta = opts_flow$get("ref_fasta"),
  java_exe = opts_flow$get("java_exe"),
  java_mem = opts_flow$get("java_mem"),
  java_tmp = opts_flow$get("java_tmp"),
  gatk_jar = opts_flow$get("gatk_jar"),
  haplotyper_opts = opts_flow$get("haplotyper_opts"),
  cpu_haplotyper = opts_flow$get("cpu_haplotyper")
)
}
\arguments{
\item{x}{input bam}

\item{haplotyper_opts}{all additional arguments supported by GATK}
}
\description{
a wrapper around gatk haplotyper
}
\details{
For more details refer to: https://www.broadinstitute.org/gatk/gatkdocs/org_broadinstitute_gatk_tools_CatVariants.php
}
