% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlavbseq.R
\name{hla_fqs}
\alias{hla_fqs}
\title{hla_fqs}
\usage{
hla_fqs(bam, samplename, samtools_exe = opts_flow$get("samtools_exe"),
  java_exe = opts_flow$get("java_exe"),
  java_mem = opts_flow$get("java_mem"),
  picard_jar = opts_flow$get("picard_jar"),
  region = "6:29690552-33111102", execute = TRUE)
}
\arguments{
\item{bam}{}

\item{samplename}{}

\item{region}{}

\item{execute}{}
}
\description{
hla_fqs
}
