% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlavbseq.R
\name{hlavbseq}
\alias{hlavbseq}
\title{runHLAVBSeq}
\usage{
hlavbseq(bam, samplename, region = "6:29690552-33111102",
  paired_end = TRUE, bwa_exe = opts_flow$get("bwa_exe"),
  hla_all_fa = opts_flow$get("hla_all_v2_fa"),
  hla_allele_list = opts_flow$get("hla_allele_list"),
  java_exe = opts_flow$get("java_exe"),
  hlavbseq_jar = opts_flow$get("hlavbseq_jar"),
  hlavbseq_call_hla_digits_py = opts_flow$get("hlavbseq_call_hla_digits_py"),
  hla_call_digits_opts = opts_flow$get("hla_call_digits_opts"))
}
\arguments{
\item{region}{}

\item{bamName}{}

\item{odir}{}

\item{oprefix}{}

\item{paired}{}
}
\description{
runHLAVBSeq
}
\details{
follow this link for more details:

http://nagasakilab.csml.org/hla/
}
