% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htseq.R
\name{htseq}
\alias{htseq}
\title{A wrapper around the RNASeq counting tool, htseq}
\usage{
htseq(bam, samplename, outprefix, htseq_exe = opts_flow$get("htseq_exe"),
  htseq_opts = opts_flow$get("htseq_opts"), gtf = opts_flow$get("gtf"))
}
\arguments{
\item{samplename}{samplename}

\item{outprefix}{output prefix to be used}

\item{htseq_opts}{options to be supplied to htseq, default retrieved using opts_flow$get("htseq_opts").}

\item{gtf}{a GTF file used}

\item{x}{a bam file}
}
\description{
A wrapper around the RNASeq counting tool, htseq
}
