% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto.R
\name{kallisto_pe}
\alias{kallisto_pe}
\title{Run a paired end kallisto pipeline}
\usage{
kallisto_pe(
  fq1,
  fq2,
  samplename,
  outprefix,
  kallisto_exe = opts_flow$get("kallisto_exe"),
  kallisto_index_path = opts_flow$get("kallisto_index_path"),
  kallisto_opts = opts_flow$get("kallisto_opts")
)
}
\arguments{
\item{fq1}{something}

\item{fq2}{something}

\item{samplename}{something}

\item{outprefix}{something}

\item{kallisto_exe}{something}

\item{kallisto_index_path}{something}

\item{kallisto_opts}{something}
}
\description{
Run a paired end kallisto pipeline
}
