% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_dnaseq.R
\name{metadata_for_dnaseq}
\alias{metadata_for_dnaseq}
\title{metadata_for_dnaseq}
\usage{
metadata_for_dnaseq(
  trk,
  normalize_file_names = F,
  basename_of_files = normalize_file_names,
  stage = "none",
  check_files_exist = F
)
}
\arguments{
\item{trk}{}

\item{normalize_file_names}{}

\item{basename_of_files}{}

\item{stage}{}

\item{check_files_exist}{}
}
\description{
metadata_for_dnaseq
}
