% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{modules}
\alias{modules}
\title{Show a list of all available modules in ultraseq}
\usage{
modules(pkg = "ultraseq", show = TRUE)
}
\arguments{
\item{pkg}{something}

\item{show}{something}
}
\description{
Show a list of all available modules in ultraseq
}
