% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpileup.R
\name{mpileup}
\alias{mpileup}
\title{Title}
\usage{
mpileup(bam, samplename = "samp",
  ref_fa = "/rsrch2/iacs/iacs_dep/sseth/ref/human/annotations/gencode/v19/GRCh37.p13.genome.fa",
  regions_bed = "/rsrch2/iacs/iacs_dep/sseth/projects2/ss_pdac_subtyping/20161012_rna_calls/regions.bed",
  outfile_raw, outfile_call, samtools_exe = "samtools",
  bcftools_exe = "bcftools", execute = FALSE)
}
\arguments{
\item{bam}{}

\item{bed}{}

\item{df_bed}{}
}
\description{
Title
}
