% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_mutect.R
\name{mutect.read}
\alias{mutect.read}
\title{read mutect}
\usage{
mutect.read(trk, col_fl = "mutect", col_samp = "samplelbl1")
}
\arguments{
\item{trk}{something}

\item{col_fl}{something}

\item{col_samp}{something}
}
\description{
read mutect
}
