% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutect1.7.R
\name{mutect1.7}
\alias{mutect1.7}
\title{A wrapper around somatic mutation caller MuTect}
\usage{
mutect1.7(
  trk,
  samplename,
  variant_calling_mode,
  java_exe = opts_flow$get("java17_exe"),
  java_tmp = opts_flow$get("java17_tmp"),
  java_mem = opts_flow$get("java17_mem_str"),
  gatk4_exe = opts_flow$get("gatk4_exe"),
  ref_fasta = opts_flow$get("ref_fasta"),
  mutect1_jar = opts_flow$get("mutect1_jar"),
  mutect1_cpu = opts_flow$get("mutect1_cpu"),
  mutect1_mem_str = opts_flow$get("mutect1_mem_str"),
  mutect1_opts = opts_flow$get("mutect1_opts"),
  mutect1_germline_vcf = opts_flow$get("mutect1_germline_vcf"),
  mutect1_cosmic_vcf = opts_flow$get("mutect1_cosmic_vcf"),
  mutect1_pon_vcf = opts_flow$get("mutect1_pon_vcf"),
  mutect1_dir = pipe_str$mutect1$dir,
  interval_split = opts_flow$get("mutect1_intervals")
)
}
\arguments{
\item{samplename}{name of the sample, to be added to the flowmat}

\item{java_exe}{path to java's executable [java]}

\item{java_tmp}{path to a temp folder to be used by java}

\item{tumor_bam}{path to a tumor bam file}

\item{normal_bam}{path to a normal bam file}

\item{outprefix}{output file name [optional].
By default, this is created using names of tumor and normal bam files.}

\item{is_merged}{specify if the input bam is already split by chromosomes (FALSE) 
or is a merged file with all chromosome (TRUE). [FALSE]}

\item{split_by_chr}{fork running mutect by chromosome to make it faster [TRUE].
Turning is option OFF is not fully supported.}

\item{mutect_jar}{path to mutect's jar file}

\item{cpu_mutect}{integer specifying number of thread to be used per mutect fork}

\item{mem_mutect}{amount of memory to be used by mutect [-Xmx12g]}

\item{ref_fasta_path}{path to the reference genome in fasta format}

\item{mutect_opts}{additional arguments passed onto mutect}
}
\value{
The function returns a flowmat with all the commands. 
The final file is called \code{'outprefix'_merged.mutect.tsv}.
}
\description{
This generates a set of commandlines, per chromosome
}
\examples{
\dontrun{

x = "tumor.bam"
y = "normal.bam"

out = mutect(x, y, samplename="tumor_normal", is_merged = TRUE)

}
}
