% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutect2.R
\name{mutect2.gatk4}
\alias{mutect2.gatk4}
\title{https://github.com/bcbio/bcbio-nextgen/issues/1969x
That does sound reasonable to me -- we tend to run WGS somatic samples at 128 or 256 cores, and the whole pipeline takes 40 or 30 hours, respectively. 
MuTect2 and VarDict can stall on regions with extremely high coverage but that usually results in no CPU activity, i.e., the process just hangs.
https://gatkforums.broadinstitute.org/gatk/discussion/4133/when-should-i-use-l-to-pass-in-a-list-of-intervals
Below is a step-by-step breakdown of the Best Practices workflow, with a detailed explanation of why -L should or shouldn’t be used with each tool.}
\usage{
mutect2.gatk4(tumor_bam, normal_bam, tumor_name, normal_name,
  samplename = opts_flow$get("samplename"), split_by_chr = TRUE,
  ref_fasta = opts_flow$get("ref_fasta"),
  java_exe = opts_flow$get("java_exe"),
  java_mem = opts_flow$get("java_mem"),
  gatk4_exe = opts_flow$get("gatk4_exe"),
  mutect.cpu = opts_flow$get("mutect.cpu"),
  mutect.mem = opts_flow$get("mutect.mem"),
  mutect_opts = opts_flow$get("mutect_opts"),
  germline_vcf = opts_flow$get("germline_vcf"),
  germline_small_vcf = opts_flow$get("germline_small_vcf"),
  gatk_intervals_fl = opts_flow$get("gatk_C1_intervals"))
}
\arguments{
\item{tumor_bam}{path to a tumor bam file}

\item{normal_bam}{path to a normal bam file}

\item{samplename}{name of the sample, to be added to the flowmat}

\item{split_by_chr}{fork running mutect by chromosome to make it faster [TRUE].
Turning is option OFF is not fully supported.}

\item{ref_fasta}{path to the reference genome in fasta format}

\item{java_exe}{path to java's executable [java]}

\item{mutect_opts}{additional arguments passed onto mutect}

\item{out_prefix}{output file name [optional].
By default, this is created using names of tumor and normal bam files.}

\item{is_merged}{specify if the input bam is already split by chromosomes (FALSE) 
or is a merged file with all chromosome (TRUE). [FALSE]}

\item{java_tmp}{path to a temp folder to be used by java}

\item{mutect_jar}{path to mutect's jar file}

\item{cpu_mutect}{integer specifying number of thread to be used per mutect fork}

\item{mem_mutect}{amount of memory to be used by mutect [-Xmx12g]}
}
\value{
The function returns a flowmat with all the commands. 
The final file is called \code{'out_prefix'_merged.mutect.tsv}.
}
\description{
Tool	-L?	Why / why not
BaseRecalibrator	YES	This excludes off-target sequences and sequences that may be poorly mapped, which have a higher error rate. Including them could lead to a skewed model and bad recalibration.
PrintReads	NO	Output is a bam file; using -L would lead to lost data.
UnifiedGenotyper/Haplotype Caller	YES	We’re only interested in making calls in exome regions; the rest is a waste of time & includes lots of false positives.
Next steps	NO	No need since subsequent steps operate on the callset, which was restricted to the exome at the calling step.
A wrapper around somatic mutation caller MuTect
}
\details{
This generates a set of commandlines, per chromosome

following this: 
https://software.broadinstitute.org/gatk/documentation/article?id=11136
https://gatkforums.broadinstitute.org/gatk/discussion/11127

param details:
https://software.broadinstitute.org/gatk/documentation/tooldocs/current/org_broadinstitute_hellbender_tools_walkers_mutect_Mutect2.php
}
\examples{
\dontrun{

x = "tumor.bam"
y = "normal.bam"

out = mutect(x, y, samplename="tumor_normal", is_merged = TRUE)

}
This produces a six-column table as shown. The alt_count is the count of reads that support the ALT allele in the germline resource. 
The allele_frequency corresponds to that given in the germline resource. Counts for other_alt_count refer to reads that support all other alleles.
Second, perform orientation bias filtering with FilterByOrientationBias. 
We provide the tool with the once-filtered calls 9_somatic_oncefiltered.vcf.gz, 
the pre_adapter_detail_metrics file and the sequencing contexts for FFPE (C→T transition) and OxoG (G→T transversion). 
The tool knows to include the reverse complement contexts
https://software.broadinstitute.org/gatk/documentation/tooldocs/current/org_broadinstitute_hellbender_tools_exome_FilterByOrientationBias.php
}
