% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gantt.R
\name{norm_offset}
\alias{norm_offset}
\title{Title}
\usage{
norm_offset(x, format = "\%Y-\%m-\%d \%H:\%M:\%S",
  origin = "1970-01-01")
}
\arguments{
\item{x}{time in character format}
}
\description{
Title
}
