% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncotator.R
\name{oncotator}
\alias{oncotator}
\title{creating a maf file
Mandatory fields: 
Hugo_Symbol, Chromosome, Start_Position, End_Position, Reference_Allele, 
Tumor_Seq_Allele2, Variant_Classification, Variant_Type and Tumor_Sample_Barcode.}
\usage{
oncotator(x, samplename, outfile, input_format = "maflite",
  output_format = "maflite",
  oncotator_exe = "~/apps/python/2.7/bin/oncotator",
  db_dir = "~/rsrch2_home/ref/human/b37/annotations/oncotator/oncotator_v1_ds_April052016",
  execute = F, oncotator_opts = "-c $tx")
}
\arguments{
\item{x}{tsv file, Hugo_Symbol, Chromosome, Start_Position, End_Position,Reference_Allele, Tumor_Seq_Allele2, OR a VCF file}

\item{outfile}{}

\item{oncotator_exe}{}

\item{db_dir}{}
}
\description{
creating a maf file
Mandatory fields: 
Hugo_Symbol, Chromosome, Start_Position, End_Position, Reference_Allele, 
Tumor_Seq_Allele2, Variant_Classification, Variant_Type and Tumor_Sample_Barcode.
}
