% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie.R
\name{parse_bowtie_metrics}
\alias{parse_bowtie_metrics}
\title{parse_bowtie_metrics}
\usage{
parse_bowtie_metrics(wd, files, aln.dir, cores = 1)
}
\arguments{
\item{wd}{working dir}

\item{files}{out files}
}
\description{
parse_bowtie_metrics
}
\examples{
parse_bowtie_report(files = files, wd = wd)
}
