% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picard_parsers.R
\name{parse_flagstat}
\alias{parse_flagstat}
\title{Title}
\usage{
parse_flagstat(x, genome_length = opts_flow$get("genome_length"),
  read_length = opts_flow$get("read_length"),
  verbose = opts_flow$get("verbose"))
}
\arguments{
\item{names.prefix}{}
}
\description{
Title
}
