% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picard_parsers.R
\name{parse_hsmetrics}
\alias{parse_hsmetrics}
\title{hsmetrics}
\usage{
parse_hsmetrics(outFile, names.prefix = "")
}
\arguments{
\item{outFile}{something}

\item{names.prefix}{something}
}
\description{
hsmetrics
}
