% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto.R
\name{parse_kallisto_metrics}
\alias{parse_kallisto_metrics}
\title{parse_kallisto_metrics}
\usage{
parse_kallisto_metrics(wd, pattern = "kallisto")
}
\arguments{
\item{wd}{path for flowr dir}

\item{pattern}{something}
}
\description{
parse_kallisto_metrics
}
