% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_vcf.R
\name{parse_somatic_vcf.oncotator}
\alias{parse_somatic_vcf.oncotator}
\title{Parse a somatic VCF, with two samples.}
\usage{
parse_somatic_vcf.oncotator(x, samp, ref)
}
\arguments{
\item{x}{a vcf file}

\item{samp}{name of the 'tumor' sample}

\item{ref}{name of the 'reference' sample}
}
\description{
Parse a somatic VCF, with two samples.
}
