% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_vcf.R
\name{parse_vcf}
\alias{parse_vcf}
\title{Parse a VCF file into a TSV to be processed and annotated}
\usage{
parse_vcf(x, cores = 1)
}
\arguments{
\item{x}{a vcf file to be parsed}

\item{cores}{something}
}
\description{
Parse a VCF file into a TSV to be processed and annotated
}
