% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picard.R
\name{picard_mergesamfiles}
\alias{picard_mergesamfiles}
\title{Use picard's MergeSamFiles tool to merge bam/sam files}
\usage{
picard_mergesamfiles(bams, outfile, samplename = opts_flow$get("samplename"),
  java_exe = opts_flow$get("java_exe"),
  java_mem = opts_flow$get("java_mem"),
  java_tmp = opts_flow$get("java_tmp"),
  picard_jar = opts_flow$get("picard_jar"),
  mergesam_opts = "ASSUME_SORTED=TRUE VALIDATION_STRINGENCY=LENIENT CREATE_INDEX=true USE_THREADING=true")
}
\arguments{
\item{bams}{a vectors of files to merge}

\item{outfile}{name of the merged bam file}

\item{samplename}{}

\item{java_exe}{path to java}

\item{java_mem}{java memory}

\item{picard_jar}{}
}
\description{
The resulting bam file is sorted and with a indexed created.
Validation stringency of inputs is kept as lenient.
Multi-threading is turned on by default, though in our experience this does
not seem to use a lot of threads.
}
\examples{
\dontrun{

picard_mergesamfiles(bams = c("a.bam", "b.bam"), "merged.bam", samplename = "ab")

}
}
