% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picard.R
\name{picard_reorder}
\alias{picard_reorder}
\title{Use picard's reorder bam file}
\usage{
picard_reorder(x, outfile, samplename = opts_flow$get("samplename"),
  java_exe = opts_flow$get("java_exe"),
  java_mem = opts_flow$get("java_mem"),
  java_tmp = opts_flow$get("java_tmp"),
  picard_dir = opts_flow$get("picard_dir"),
  picard_reorder_opts = opts_flow$get("picard_reorder_opts"),
  ref_fasta = opts_flow$get("ref_fasta"),
  samtools_exe = opts_flow$get("samtools_exe"))
}
\arguments{
\item{x}{a vectors of files to merge}

\item{samplename}{something}

\item{java_exe}{something}

\item{java_mem}{something}

\item{java_tmp}{something}

\item{picard_dir}{something}

\item{mergedbam}{something}
}
\description{
The resulting file is sorted and index is created for it.
Validation stringency of inputs is kept as lenient.
Multi-threading is turned on by default, though in our experience this does
not seem to use a lot of threads.
}
