% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picard.R
\name{picard_rg}
\alias{picard_rg}
\title{picard_rg}
\usage{
picard_rg(x, samplename = opts_flow$get("samplename"),
  rg_lane = opts_flow$get("rg_lane"),
  rg_platform = opts_flow$get("rg_platform"),
  rg_center = opts_flow$get("rg_center"),
  java_exe = opts_flow$get("java_exe"),
  java_mem = opts_flow$get("java_mem"),
  java_tmp = opts_flow$get("java_tmp"),
  picard_jar = opts_flow$get("picard_jar"))
}
\description{
picard_rg
}
