% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picard.R
\name{picard_samtofastq}
\alias{picard_samtofastq}
\title{Title}
\usage{
picard_samtofastq(bam, samplename = opts_flow$get("samplename"),
  paired = TRUE, split = FALSE, num_reads = 8e+06,
  split_fq_exe = system.file("scripts/split_fq", package = "ngsflows"),
  java_exe = opts_flow$get("java_exe"),
  java_mem = opts_flow$get("java_mem"),
  java_tmp = opts_flow$get("java_tmp"),
  picard_dir = opts_flow$get("picard_dir"),
  bam_fastq_opts = "INCLUDE_NON_PF_READS=true INCLUDE_NON_PRIMARY_ALIGNMENTS=false RE_REVERSE=true VALIDATION_STRINGENCY=LENIENT")
}
\arguments{
\item{picard_dir}{}
}
\description{
Title
}
