% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk_preprocess.R
\name{preprocess.gatk_v4}
\alias{preprocess.gatk_v4}
\title{Pre-process bam files following Broad's best practices for variant calling, starting from aligned BAM file}
\usage{
preprocess.gatk_v4(bam, outfile,
  samplename = opts_flow$get("samplename"),
  split_by_chr = opts_flow$get("split_by_chr"),
  samtools_exe = opts_flow$get("samtools.exe"),
  java_exe = opts_flow$get("java_exe"),
  java_mem = opts_flow$get("java_mem_str"),
  ref_fasta = opts_flow$get("ref_fasta"),
  gatk_jar = opts_flow$get("gatk4.jar"),
  gatk_opts = opts_flow$get("gatk4.opts"),
  markdup_mem = opts_flow$get("gatk4.markdup.mem"),
  markdup_opts = opts_flow$get("gatk4.markdup.opts"),
  baserecalib_mem = opts_flow$get("gatk4.baserecalib.mem"),
  baserecalib_opts = opts_flow$get("gatk4.baserecalib.opts"),
  execute_cmds = FALSE)
}
\arguments{
\item{bam}{bam file path}

\item{outfile}{output file name}

\item{samplename}{name of the sample}

\item{split_by_chr}{split processing by chromosomr where ever possible}

\item{samtools_exe}{path to samtools}

\item{java_exe}{path to java}

\item{execute_cmds}{run commands, after creation. Useful for testing/debugging and running on local platforms.}

\item{java_tmp}{path to java_tmp, can leave blank}

\item{gatk_jar_path}{path to gatk jar file}

\item{picard_jar_path}{path to picard jar file}

\item{ref_fasta_path}{reference fasta file}

\item{picard_markdup_opts}{a character vector of options for picard mark duplication step}

\item{gatk_target_opts}{a character vector of options for gatk target step}

\item{gatk_realign_opts}{a character vector of options for gatk realign step}

\item{gatk_baserecalib_opts}{a character vector of options for gatk baserecalib step}

\item{gatk_printreads_opts}{a character vector of options for gatk printreads step}

\item{mem_markdup}{memory used by java, example -Xmx1g}

\item{mem_target}{memory used by java, example -Xmx1g}

\item{mem_realign}{memory used by java, example -Xmx1g}

\item{mem_baserecalib}{memory used by java, example -Xmx1g}

\item{mem_printreads}{memory used by java, example -Xmx1g}

\item{cpu_markdup}{not used.}

\item{cpu_target}{number of threads used for GATK target creation step}

\item{cpu_realign}{number of cpu used}

\item{cpu_baserecalib}{number of cpu used}

\item{cpu_printreads}{number of cpu used}
}
\description{
This function provides a wrapper around the best practices described on 
\href{https://software.broadinstitute.org/gatk/best-practices/workflow?id=11165}{GATK's website}.
If the link is broken, google 'GATK best practices'

This aims to perform the following steps ( for DNA ):

\itemize{
\item mark duplicates
\item realign indels
\item recalibrate bases
\item current version: \emph{3.4-46}
}

For RNA GATK recommends a additional step of split n trim, which is not currently supported (contributions welcome !).

\strong{NOTE}:

Some GATK tools use \href{https://www.broadinstitute.org/gatk/guide/article?id=1975}{CPU threads while others use data threads},
flowr tries to use efficiently make the best use of both/either depending on tool's compatibility.

1.   call MarkDuplicates ----> SortAndFixTags
2. Perform Base Quality Score Recalibration (BQSR) on the sorted BAM in parallel (BaseRecalibrator)
}
\details{
Flow following Broad's best practices for variant calling, starting from sorted bam
}
\examples{
\dontrun{
## load options, including paths to tools and other parameters
opts_flow$load(flowr::fetch_conf("ultraseq.conf"), check = FALSE)
out = preprocess("my_wex.bam", samplename = "samp", split_by_chr = TRUE)

}
}
