% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyclone.R
\name{pyclone.run_analysis_pipeline}
\alias{pyclone.run_analysis_pipeline}
\title{run pyclone cmd}
\usage{
pyclone.run_analysis_pipeline(df, odir,
  pyclone_params = "--num_iters 10000 --density pyclone_beta_binomial --prior parental_copy_number --burnin 2500")
}
\arguments{
\item{df}{}

\item{odir}{}

\item{pyclone_params}{}
}
\description{
run pyclone cmd
}
\details{
--num_iters 10000, usually, increase to 50000
}
