% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citup.R
\name{pyclone_citup}
\alias{pyclone_citup}
\title{pyclone_citup}
\usage{
pyclone_citup(
  pyclone_path,
  citup_exe = "module load conda_/2.7;source activate citup;run_citup_qip.py",
  citup_params = "--submit local --loglevel DEBUG --maxjobs 8"
)
}
\arguments{
\item{pyclone_path}{}

\item{citup_exe}{}

\item{citup_params}{}
}
\description{
this step would use 16 cores
}
