% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyclone.R
\name{pyclone_flo}
\alias{pyclone_flo}
\title{funcs, ending with fl, denote output is a flowmat type object}
\usage{
pyclone_flo(pyclone_input,
  pyclone_exe = "module load conda_/2.7;source activate pyclone;PyClone",
  samplename, out_prefix)
}
\description{
funcs, ending with fl, denote output is a flowmat type object
}
