% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyclonevi.R
\name{pyclone_pipe_v1_purecn_flo}
\alias{pyclone_pipe_v1_purecn_flo}
\title{this function is invoked from ss_het
creates all the commands to run pyclone and subsequent steps}
\usage{
pyclone_pipe_v1_purecn_flo(
  trk,
  samplename,
  odir,
  flow_conf,
  bam_readcount_r_func = bam_readcount_r.ssm,
  fetch_files.exe = opts_flow$get("fetch_files.exe")
)
}
\description{
this function is invoked from ss_het
creates all the commands to run pyclone and subsequent steps
}
