% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyclone_pipe_v1.R
\name{pyclone_pipe_v1_step2_r}
\alias{pyclone_pipe_v1_step2_r}
\title{pyclone_pipe_v2_step1_r}
\usage{
pyclone_pipe_v1_step2_r(pyclone_path = ".", trk = "trk.tsv",
  conf = "~/Dropbox/public/flow-r/my.ultraseq/pipelines/dnaseq/facets_pyclone/facets_pyclone_citup_v1.conf",
  execute = F)
}
\arguments{
\item{pyclone_path}{tmp, path with all relavent files}

\item{trk}{trk}

\item{conf}{conf}
}
\description{
filter bamreadcount, and run pyclone
}
