% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyclone_pipe_v1.R
\name{pyclone_pipe_v1_step3_r}
\alias{pyclone_pipe_v1_step3_r}
\title{bam_readcount_pyclone_r_step3}
\usage{
pyclone_pipe_v1_step3_r(
  pyclone_path = ".",
 
    conf = "~/Dropbox/public/flow-r/my.ultraseq/pipelines/dnaseq/facets_pyclone/facets_pyclone_citup_v1.conf"
)
}
\arguments{
\item{pyclone_path}{pyclone_path}

\item{conf}{conf}
}
\description{
pyclone plots citup/cloneevol
}
