% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyclone.R
\name{pyclone_prep_input}
\alias{pyclone_prep_input}
\title{prepare input for pyclone}
\usage{
pyclone_prep_input(segfl, mutfl, lst_bams = list(),
  segfl_type = c("sequenza", "facets"), mutfl_type = c("mutect_ann",
  "mutect_bamreadcount"), outfile)
}
\arguments{
\item{seg}{sequenza segment file}

\item{mut}{mutect mutations}
}
\description{
prepare input for pyclone
}
\details{
need to expand to support facets input
}
