% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_htseq.R
\name{read_htseq}
\alias{read_htseq}
\title{Reads multiple HTSeq/QORTS files and creates a simple data.frame of all the counts.}
\usage{
read_htseq(x, y, num_cores = 10)
}
\arguments{
\item{x}{a vector of files}

\item{y}{a vector of sample names}
}
\description{
Reads multiple HTSeq/QORTS files and creates a simple data.frame of all the counts.
}
