% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvacseq.R
\name{read_pvacseq2}
\alias{read_pvacseq2}
\title{Title}
\usage{
read_pvacseq2(oprefix, seq_sample_id, path_patient_id)
}
\arguments{
\item{x}{}
}
\description{
Title
}
\details{
The pVACseq pipeline will write its results in separate folders depending on which prediction algorithms were chosen:
MHC_Class_I: for MHC class I prediction algorithms
MHC_Class_II: for MHC class II prediction algorithms
combined: If both MHC class I and MHC class II prediction algorithms were run, this folder combines the neoeptiope predictions from both

Each folder will contain the same list of output files (listed in the order created):
  File Name Description <sample_name>.tsv
  An intermediate file with variant, transcript, coverage, vaf, and expression information parsed from the input files.
  <sample_name>.tsv_<chunks> (multiple
  The above file but split into smaller chunks for easier processing with IEDB.
  
  <sample_name>.all_epitopes.tsv
  A list of all predicted epitopes and their binding affinity scores, with additional variant information from the <sample_name>.tsv.
  <sample_name>.filtered.tsv 
  The above file after applying all filters, with cleavage site and stability predictions added.
  <sample_name>.filtered.condensed.ranked.tsv
  A condensed version of the filtered TSV with only the most important columns remaining, with a priority score for each neoepitope candidate added.
}
