% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{read_theta.results}
\alias{read_theta.results}
\title{to_theta_cn_file.gatkcnv}
\usage{
read_theta.results(oprefix, path)
}
\arguments{
\item{oprefix}{expecting files named `{path}/{oprefix}.input"`,  `{path}/{oprefix}.n2.results"` etc.}

\item{path}{output dir, can be ".", or ""}
}
\description{
to_theta_cn_file.gatkcnv
}
