% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parse.R
\name{read_vcf_somatic}
\alias{read_vcf_somatic}
\title{Parse a somatic VCF, with two samples.}
\usage{
read_vcf_somatic(x, samp = NULL, ref = NULL)
}
\arguments{
\item{x}{a vcf file}

\item{samp}{name of the 'tumor' sample}

\item{ref}{name of the 'reference' sample}
}
\description{
Parse a somatic VCF, with two samples.
}
