% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_sh_sg.R
\name{run_blast_map_seq}
\alias{run_blast_map_seq}
\title{Title}
\usage{
run_blast_map_seq(mp,
  blast_db = "/rsrch2/iacs/iacs_dep/sseth/ref/human/annotations/gencode/v19/gencode.v19.pc_transcripts.fa",
  out_prefix, do_revcomp = TRUE)
}
\arguments{
\item{mp}{mapping file with columns: antisense_seq sense_seq}

\item{blast_db}{}

\item{out_prefix}{}

\item{do_revcomp}{}
}
\description{
Title
}
