% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cmd.R
\name{run_cmd}
\alias{run_cmd}
\title{run_cmd}
\usage{
run_cmd(cmd, cmdname, target, stderr = "/dev/stderr", redo = F)
}
\arguments{
\item{cmd}{cmd to run}

\item{cmdname}{a descriptive name of the cmd}

\item{target}{target file, cmd is skipped if present}

\item{stderr}{redirect output (stderr, 2>) to a file}

\item{redo}{force rede}
}
\description{
run_cmd
}
