% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenza.R
\name{sequenza_r_pipe}
\alias{sequenza_r_pipe}
\title{window: ONLY for plotting, window size controlled by breaks.method ONLY
https://groups.google.com/forum/#!searchin/sequenza-user-group/too$20many$20segments%7Csort:date/sequenza-user-group/Hnzswcq-34s/nZ94ehwAGQAJ
min.type.freq=0.9: to remove freq type (in case of multiple muts, at location)
    var ignored at: 5/6(0.83) 0/6(0) and 1/6(0.17)
weighted.mean=TRUE; eighted mean, positions with more reads have more weight 
    in the calculation of the mean value for the segment 
    (a segment value is an average of all the positions included in it)... 
    In some case is better with the feature turned off in presence of bias 
    (positions with huge amount of reads)}
\usage{
sequenza_r_pipe(seqz, oprefix, odir, is_male = TRUE, chroms = c(1:22,
  "X"), cores = 10, ....)
}
\description{
N.ratio.filter=10; minimum no of obs, required to consider a segmetn
The fit step is particularly important to filter out unreliable segments. 
This will produce the estimation, we are not throwing out segments for 
downstream analysis.
}
\details{
N.BAF.filter=1; ONLY HETEROZYGOUS SNPS ARE USED
  So a segment could have 100 "ratio" pos, but only 2 "heterozygous" pos.
  I think you are confusing yourself with the variants in the segment in this case: 
  the only variant that we consider for the segment are the germlines heterozygous, 
  so position that are detected het already in the normal. 
  
ratio.priority: FALSE
 This is advisable if the BAF profile looks very noisy. 
 In WES it could be the case, in WGS the heterozygous position are much more and more 
 reliable, but in exome it depends a lot on the specific dataset.
}
