% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_fq.R
\name{split_fq}
\alias{split_fq}
\title{split fastq into pre-determined number of splits}
\usage{
split_fq(x, samplename, jobname,
  split_fq_exe = system.file("scripts/split_fq.sh", package = "ultraseq"),
  chunks = 200)
}
\arguments{
\item{x}{file to split}

\item{chunks}{approx size of resulting file, in MB}
}
\description{
This calls a internal module
}
\details{
estimate the number of splits required
pre-compute the number of splits it would generate
make commands to make the split
approx, make 200mb files, for splitting...
}
