% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star.R
\name{star.2.4.2a}
\alias{star.2.4.2a}
\alias{star}
\title{A flowr wrapper for STAR aligner}
\usage{
star.2.4.2a

star.2.4.2a

star(fq1, fq2, bam,
 samplename = "samp1", out_path, out_prefix,
 ref_fasta = opts_flow$get("ref_fasta"), ref_star = opts_flow$get("ref_star"),
 star_exe = opts_flow$get("star_exe"), star_opts = opts_flow$get("star_opts"),
 star_cpu = opts_flow$get("star_cpu"))
}
\arguments{
\item{fq1}{single fastq file}

\item{fq2}{second read fastq file}

\item{bam}{alternatively use of bam as input, instead of fqs}

\item{samplename}{name of the sample}

\item{out_path}{output path}

\item{out_prefix}{output prefix}

\item{ref_fasta}{path to reference genome in fasta format}

\item{ref_star}{path to star's reference index}

\item{star_exe}{path to star executable}

\item{star_opts}{additional options passed onto star. This is a simple character vector appended to the commandline.}

\item{star_cpu}{a integer specifying the number of cores to be used}
}
\description{
For more details refer to \href{https://github.com/alexdobin/STAR/raw/master/doc/STARmanual.pdf}{STAR's manual}.
Starting 2.4.2a, star supports a twoPass mode, thus simplifying this pipeline.
}
