% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superfreq.R
\name{superfreq}
\alias{superfreq}
\title{superfreq}
\usage{
superfreq(odir = ".", ...)
}
\arguments{
\item{odir}{dir to store final results}

\item{...}{all params passed onto superFreq::superFreq}
}
\description{
call superFreq from the pkg and store the results
}
