% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titancna.R
\name{titancna_somatic_matched}
\alias{titancna_somatic_matched}
\title{Title}
\usage{
titancna_somatic_matched(
  trk,
  samplename,
  run_cmds = F,
  funr_exe = opts_flow$get("R_funr_exe"),
  num_cores = 4,
  rscript_exe = opts_flow$envir$R35_script,
  titancna_dir = pipe_str$titancna$dir,
  titancna_exe = opts_flow$envir$titancna_exe,
  titancna_centromere_gap_fl = opts_flow$envir$titancna_centromere_gap_fl,
  titancna_cluster_min = opts_flow$envir$titancna_cluster_min,
  titancna_cluster_max = opts_flow$envir$titancna_cluster_max,
  titancna_opts = opts_flow$envir$titancna_opts
)
}
\arguments{
\item{trk}{}

\item{samplename}{}

\item{num_cores}{}

\item{test}{}
}
\description{
Title
}
