% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlavbseq.R
\name{to_hla_pvacseq.hlavbseq}
\alias{to_hla_pvacseq.hlavbseq}
\title{to_hla_pvacseq.hlavbseq}
\usage{
to_hla_pvacseq.hlavbseq(x, hla_fl,
  allele_db_fl = "~/.rsrch2/iacs/iacs_dep/sseth/ref/human/b37/annotations/hlavbseq/all_alleles.txt")
}
\arguments{
\item{x}{d4 file}

\item{hla_fl}{}
}
\description{
to_hla_pvacseq.hlavbseq
}
