% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_mae.seg.R
\name{to_long.gr_seg}
\alias{to_long.gr_seg}
\title{to_long.igv_seg}
\usage{
to_long.gr_seg(gr_seg, expand_using = c("marks", "genes"))
}
\arguments{
\item{expand_using}{}
}
\description{
convert the IGV segments to genes (expanded version)
}
