% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlavbseq.R
\name{to_pvacseq.hlavbseq}
\alias{to_pvacseq.hlavbseq}
\title{to_hla_pvacseq.hlavbseq}
\usage{
to_pvacseq.hlavbseq(
  x,
  outfile1,
  outfile2,
  allele_db_fl = opts_flow$get("hla_allele_db")
)
}
\arguments{
\item{x}{d4 file}

\item{outfile1}{}
}
\description{
to_hla_pvacseq.hlavbseq
}
\details{
more details here:
http://nagasakilab.csml.org/hla/
}
