% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optitype.R
\name{to_pvacseq.optitype}
\alias{to_pvacseq.optitype}
\title{to_pvacseq.optitype}
\usage{
to_pvacseq.optitype(path, outfile)
}
\arguments{
\item{path}{}

\item{outfile}{outfile}
}
\description{
convert optitype for pvacseq
}
