% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phlat.R
\name{to_pvacseq.phlat}
\alias{to_pvacseq.phlat}
\title{to_pvacseq.phlat}
\usage{
to_pvacseq.phlat(phlat_fl, outfile1, outfile2)
}
\arguments{
\item{outfile1}{outfile}

\item{outfile2}{outfile}

\item{path}{}
}
\description{
convert optitype for pvacseq
}
