% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polysolver.R
\name{to_pvacseq.polysolver}
\alias{to_pvacseq.polysolver}
\title{to_pvacseq.polysolver}
\usage{
to_pvacseq.polysolver(x, outfile)
}
\arguments{
\item{outfile}{outfile}

\item{path}{}
}
\description{
convert optitype for pvacseq
}
