% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{to_theta_cn_file.gatkcnv}
\alias{to_theta_cn_file.gatkcnv}
\title{to_theta_cn_file.gatkcnv}
\usage{
to_theta_cn_file.gatkcnv(cnr_file, gatkcnv_cnts, theta_cnr_file)
}
\arguments{
\item{titan_hets_file}{}
}
\description{
to_theta_cn_file.gatkcnv
}
