% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{to_theta_het_file.gatkcnv}
\alias{to_theta_het_file.gatkcnv}
\title{to_theta_het_file.gatkcnv}
\usage{
to_theta_het_file.gatkcnv(hets_file, theta_hets_file)
}
\arguments{
\item{titan_hets_file}{}
}
\description{
to_theta_het_file.gatkcnv
}
