% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titancna.R
\name{to_titan_cn_file}
\alias{to_titan_cn_file}
\title{.titan_cn_file}
\usage{
to_titan_cn_file(cnr_file, titan_cnr_file, file_type = c("gatkcnv", "cnvkit"))
}
\arguments{
\item{cnr_file}{}
}
\description{
Convert CNVkit or GATK4 normalized input into TitanCNA ready format.
}
