% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titancna.R
\name{to_titan_hets_file}
\alias{to_titan_hets_file}
\title{to_titan_hets_file}
\usage{
to_titan_hets_file(hets_file, titan_hets_file)
}
\arguments{
\item{hets_file}{}

\item{titan_hets_file}{}
}
\description{
to_titan_hets_file
}
