% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_vcf.R
\name{to_vcf.mutect1.7_cgl_ann_v2}
\alias{to_vcf.mutect1.7_cgl_ann_v2}
\title{to_vcf.mutect1.7_extended_tsv}
\usage{
to_vcf.mutect1.7_cgl_ann_v2(x, outfile, tumor_name, normal_name, verbose = T)
}
\arguments{
\item{x}{}

\item{outfile}{ouput VCF file}
}
\description{
Assumes fixed coloumns of VCF
}
\details{
this one lacks strand etc and chr is named differently
}
