% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_vcf.R
\name{to_vcf.mutect_call_stats}
\alias{to_vcf.mutect_call_stats}
\title{to_vcf.mutect_call_stats}
\usage{
to_vcf.mutect_call_stats(x, outfile)
}
\arguments{
\item{x}{}

\item{outfile}{ouput VCF file}
}
\description{
Assumes fixed coloumns of VCF
}
