% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_readcount.R
\name{bam_readcount}
\alias{bam_readcount}
\title{Run BAM readcount}
\usage{
bam_readcount(bam, samplename, bed, outfile,
  fa_fl = "~/ref/human/b37/fastas/Homo_sapiens_assembly19.fasta",
  bamreadcount_exe = "bam-readcount")
}
\arguments{
\item{bam}{}

\item{bed}{}

\item{bamreadcount_exe}{}
}
\description{
Run BAM readcount
}
