% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_set.R
\name{bam_set}
\alias{bam_set}
\title{bam_set}
\usage{
bam_set(
  bam,
  outprefix,
  ref_fasta,
  interval_split,
  split_by = c("none", "chr_len", "interval_split")
)
}
\arguments{
\item{bam}{}

\item{outprefix}{}

\item{ref_fasta}{}

\item{interval_split}{}

\item{split_by}{}
}
\value{

}
\description{
bam_set
}
