% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc.R
\name{fastqc}
\alias{fastqc}
\title{Fastqc}
\usage{
fastqc(fqs, samplename = opts_flow$get("samplename"), fqpath, odir,
  fastqc_exe = opts_flow$get("fastqc_exe"),
  cpu_fastqc = opts_flow$get("cpu_fastqc"), casava = TRUE,
  fastqc_opts = opts_flow$get("fastqc_opts"))
}
\arguments{
\item{fastqc_opts}{}
}
\description{
Fastqc
}
