% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlavbseq.R
\name{hlavbseq}
\alias{hlavbseq}
\title{runHLAVBSeq}
\usage{
hlavbseq(
  bam,
  samplename,
  paired_end = TRUE,
  odir = "hlavbseq",
  fqs,
  seq_sample_id,
  bwa_exe = opts_flow$get("bwa.exe"),
  hla_all_fa = opts_flow$get("hlavbseq.hla_v2_fa"),
  hla_allele_list = opts_flow$get("hlavbseq.hla_allele_list"),
  hla_allele_db = opts_flow$get("hlavbseq.allele_db"),
  java_exe = opts_flow$get("java.exe"),
  hlavbseq_jar = opts_flow$get("hlavbseq.jar"),
  hlavbseq_call_digits_py = opts_flow$get("hlavbseq.call_digits_py"),
  hla_call_digits_opts = opts_flow$get("hlavbseq.call_digits_opts"),
  funr_exe = opts_flow$get("R.funr_exe")
)
}
\arguments{
\item{odir}{}

\item{bamName}{}

\item{oprefix}{}

\item{region}{}

\item{paired}{}
}
\description{
runHLAVBSeq
}
\details{
follow this link for more details:

http://nagasakilab.csml.org/hla/
}
