% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlavbseq.R
\name{hlavbseq}
\alias{hlavbseq}
\title{runHLAVBSeq}
\usage{
hlavbseq(bam, samplename, region = "6:29690552-33111102",
  paired = TRUE)
}
\arguments{
\item{region}{}

\item{paired}{}

\item{bamName}{}

\item{odir}{}

\item{oprefix}{}
}
\description{
runHLAVBSeq
}
\details{
follow this link for more details:

http://nagasakilab.csml.org/hla/
}
