% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto.R
\name{kallisto_pe.fqsheet}
\alias{kallisto_pe.fqsheet}
\title{cd ~/rsrch2_home/ref/human/b37/indexes/kallisto/0.43.0/
kallisto=/rsrch2/iacs/iacs_dep/sseth/apps/kallisto/0.43.0/kallisto
$kallisto index -i gencode.v19.pc_transcripts ../../../annotations/gencode/v19/gencode.v19.pc_transcripts.fa}
\usage{
kallisto_pe.fqsheet(fqsheet,
  kallisto_exe = "/rsrch2/iacs/apps/conda/3.18.9/bin/kallisto",
  kallisto_index_path = "/rsrch2/iacs/iacs_dep/sseth/ref/human/indexes/kallisto/0.42.4/gencode.v19.pc_transcripts")
}
\description{
uses a fqsheet, which need columsn
- samplename
- lane
- num
- read
}
