% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto.R
\name{kallisto_pe}
\alias{kallisto_pe}
\title{Run a paired end kallisto pipeline}
\usage{
kallisto_pe(fq1, fq2, samplename, out_prefix,
  kallisto_exe = opts_flow$get("kallisto_exe"),
  kallisto_index_path = opts_flow$get("kallisto_index_path"),
  kallisto_opts = opts_flow$get("kallisto_opts"))
}
\arguments{
\item{fq1}{}

\item{fq2}{}

\item{samplename}{}

\item{out_prefix}{}

\item{kallisto_exe}{}

\item{kallisto_index_path}{}

\item{kallisto_opts}{}
}
\description{
Run a paired end kallisto pipeline
}
