% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titancna.R
\name{titancna_somatic_matched}
\alias{titancna_somatic_matched}
\title{Title}
\usage{
titancna_somatic_matched(trk, samplename, num_cores = nrow(df_trk), test = T)
}
\arguments{
\item{trk}{}

\item{samplename}{}

\item{num_cores}{}

\item{test}{}
}
\description{
Title
}
